Split out ifft into a file
